Move MatMulFunctor into an own matmul.h file
